Rename fetchFullCase to FetchCase
