+++ packages/api/src/services/gateways/dataStoreGateways/postgres.ts
@@ -5,7 +5,7 @@ import type DataStoreGateway from "../interfaces/dataStoreGateway"
 
 import postgresFactory from "../../db/postgresFactory"
 import caseCanBeResubmitted from "./postgres/cases/canCaseBeResubmitted"
-import fetchFullCase from "./postgres/cases/fetchFullCase"
+import fetchCase from "./postgres/cases/fetchCase"
 import fetchUserByUsername from "./postgres/users/fetchUserByUsername"
 
 class Postgres implements DataStoreGateway {
@@ -15,8 +15,8 @@ class Postgres implements DataStoreGateway {
     return await caseCanBeResubmitted(this.db, username, caseId, forceIds)
   }
 
-  async fetchFullCase(caseId: number, forceIds: number[]): Promise<Case> {
-    return await fetchFullCase(this.db, caseId, forceIds)
+  async fetchCase(caseId: number, forceIds: number[]): Promise<Case> {
+    return await fetchCase(this.db, caseId, forceIds)
   }
 
   async fetchUserByUsername(username: string): Promise<User> {
